ocr/requirements.txt
Flask==1.1.1
Flask_Cors==3.0.8
pytesseract==0.3.1
numpy==1.17.2
pillow==6.2.0
